Move SwitchToMainThreadAsync
Retrieving two  STA-services in a row on an background thread requires
at least 4 thread switches back and forth from the thread pool to the
UI thread. Given these services are already likely created by the time
this package is loaded, may as well stay on the UI thread as they will
likely just be synchronous calls.